fix(staking): refactor stake pool confirmation into multiple files
