+++ packages/staking/src/features/drawer/confirmation/StakePoolConfirmationBody/StakePoolConfirmationBody.module.scss
@@ -1,0 +1,99 @@
+@import '../../../../../../../packages/common/src/ui/styles/theme.scss';
+@import '../../../../../../../packages/common/src/ui/styles/abstracts/typography';
+
+.body {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  margin-top: size_unit(6);
+  .item {
+    align-items: center;
+    background: var(--light-mode-light-grey, var(--dark-mode-mid-grey));
+    border-radius: size_unit(2);
+    display: flex;
+    justify-content: space-between;
+    height: size_unit(13);
+    padding: 0px size_unit(4);
+
+    & + .item {
+      margin-top: size_unit(2);
+    }
+
+    &:first-child {
+      .itemData {
+        gap: 4px;
+      }
+    }
+
+    &:last-child {
+      .itemData {
+        width: 100%;
+        overflow: hidden;
+
+        div {
+          justify-content: flex-end;
+        }
+      }
+    }
+
+    & > * {
+      display: flex;
+      flex: 1 1 0;
+      min-width: 0;
+      width: 100%;
+
+      &:last-child {
+        justify-content: flex-end;
+        text-align: end;
+
+        img {
+          background: var(--color-white, #ffffff);
+          border-radius: 50%;
+          border: 0.5px solid var(--light-mode-dark-grey);
+          padding: size_unit(0.5);
+        }
+      }
+    }
+
+    .itemData {
+      color: var(--text-color-secondary);
+      display: flex;
+      flex-direction: column;
+      font-size: var(--body);
+      font-weight: 500;
+      line-height: size_unit(3);
+      width: 100%;
+
+      p {
+        margin-bottom: 0px !important;
+      }
+
+      > * {
+        width: 100%;
+      }
+
+      .dataTitle {
+        color: var(--text-color-primary);
+        font-size: var(--bodyLarge);
+        font-weight: 600;
+        line-height: size_unit(3);
+        overflow: hidden;
+        text-overflow: ellipsis;
+        white-space: nowrap;
+      }
+
+      .dataSubTitle {
+        overflow: hidden;
+        text-overflow: ellipsis;
+        white-space: nowrap;
+      }
+    }
+
+    img {
+      height: size_unit(6);
+      margin-right: size_unit(3);
+      width: size_unit(6);
+      border-radius: size_unit(10);
+    }
+  }
+}
