+++ packages/staking/src/features/drawer/confirmation/StakePoolConfirmationBody/cardano-blue.png
