+++ packages/staking/src/features/drawer/confirmation/StakePoolConfirmationBody/index.ts
@@ -1,0 +1,1 @@
+export { StakePoolConfirmationBody } from './StakePoolConfirmationBody';
