+++ packages/staking/src/features/drawer/confirmation/StakePoolConfirmationContent.module.scss
@@ -1,0 +1,50 @@
+@import '../../../../../../packages/common/src/ui/styles/theme.scss';
+@import '../../../../../../packages/common/src/ui/styles/abstracts/typography';
+
+.container {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+}
+
+.header {
+  .title {
+    align-items: center;
+    color: var(--text-color-primary);
+    display: flex;
+    font-size: var(--heading);
+    font-weight: bold;
+    line-height: size_unit(4);
+  }
+  .subTitle {
+    color: var(--text-color-secondary);
+    font-size: var(--bodyLarge);
+    font-weight: 500;
+    line-height: size_unit(3);
+    margin-top: size_unit(2);
+  }
+}
+
+.txSummaryContainer {
+  display: flex;
+  flex-direction: column;
+  gap: size_unit(3);
+  align-items: flex-start;
+
+  @media (max-width: $breakpoint-popup) {
+    padding-bottom: size_unit(7);
+  }
+}
+
+.txSummaryTitle {
+  @include text-subHeading-bold;
+  margin-top: size_unit(4);
+  margin-bottom: size_unit(4);
+  color: var(--text-color-primary);
+
+  @media (max-width: $breakpoint-popup) {
+    @include text-bodyLarge-bold;
+    margin-top: size_unit(4);
+    margin-bottom: size_unit(3);
+  }
+}
