+++ packages/staking/src/features/drawer/confirmation/StakePoolDepositReclaimDetails.module.scss
@@ -1,0 +1,52 @@
+@import '../../../../../../packages/common/src/ui/styles/theme.scss';
+@import '../../../../../../packages/common/src/ui/styles/abstracts/typography';
+
+.divider {
+  height: 1px;
+  min-height: 1px;
+  width: 100%;
+  margin: size_unit(3) 0;
+  background-color: var(--light-mode-light-grey-plus, var(--dark-mode-mid-grey));
+}
+
+.totalCostContainer {
+  margin-top: size_unit(1);
+  padding-bottom: size_unit(2);
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+
+  @media (max-width: $breakpoint-popup) {
+    margin-top: size_unit(4);
+    margin-bottom: size_unit(3);
+  }
+}
+
+.totalCostTitle {
+  @include text-subHeading-bold;
+  color: var(--text-color-primary);
+}
+
+.txSummaryContainer {
+  display: flex;
+  flex-direction: column;
+  gap: size_unit(3);
+  align-items: flex-start;
+
+  @media (max-width: $breakpoint-popup) {
+    padding-bottom: size_unit(7);
+  }
+}
+
+.txSummaryTitle {
+  @include text-subHeading-bold;
+  margin-top: size_unit(4);
+  margin-bottom: size_unit(4);
+  color: var(--text-color-primary);
+
+  @media (max-width: $breakpoint-popup) {
+    @include text-bodyLarge-bold;
+    margin-top: size_unit(4);
+    margin-bottom: size_unit(3);
+  }
+}
