+++ packages/staking/src/features/drawer/confirmation/index.ts
@@ -1,0 +1,2 @@
+export { StakePoolConfirmationContent } from './StakePoolConfirmationContent';
+export { StakePoolConfirmationFooter } from './StakePoolConfirmationFooter';
